requirements.txt
asgiref==3.6.0
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-filter==23.2
django-jazzmin==2.6.0
django-js-asset==2.0.0
djangorestframework==3.14.0
netifaces==0.11.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
